pyproject.toml
[project]
name = "ankiagent"
version = "0.1.0"
description = "AI-powered Anki flashcard creator for language learning"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-ankiconnect",
    "click",
    "httpx",
    "pillow",
    "beautifulsoup4",
    "duckduckgo-search",
    "anthropic",
]

[project.scripts]
ankiagent = "anki_agent.agents.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anki_agent"]

[dependency-groups]
dev = [
    "pytest",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "F",     # pyflakes
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
quote-style = "double"
docstring-code-line-length = 88
